fix (redis): remove redis/cache package, unused code
also, the current implementation of redis service and its integration in schedulers is wrong, we have a defined a redis service, but no endpoint to update the streams, instead we are initialzing the redis in individual schedulers, and updatig the streams there. that is each scheduler pushes to different instance of redis. our aim is single redis, in which all schedulers push and pull from, liek a central cache, following commits will rectify this